+++ src/Exceptions/RateLimitExceededException.php
@@ -7,9 +7,8 @@ declare(strict_types=1);
  *
  * (c) Eric Sizemore <admin@secondversion.com>
  *
- * This source file is subject to the MIT license. For the full
- * copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
+ * This source file is subject to the MIT license. For the full copyright and license
+ * information, please view the LICENSE file that was distributed with this source code.
  */
 
 namespace Esi\Api\Exceptions;